safari-534.53-branch/LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.4_Comments/S7.4_A4_T1-expected.txt

CONSOLE MESSAGE: line 79: SyntaxError: Parse error
S7.4_A4_T1

PASS Expected parsing failure

TEST COMPLETE
